Don't execute test_submodule_update_force on older git
Test requires the "git -C directory_name rm" command, and that is not
available in git 1.7.10 and older.

+++ src/test/java/org/jenkinsci/plugins/gitclient/GitAPITestCase.java
@@ -2031,6 +2031,10 @@ public abstract class GitAPITestCase extends TestCase {
         assertTrue("modules/ntp/CHANGELOG does not exist", w.exists("modules/ntp/CHANGELOG"));
 
         // delete the CHANGELOG file from the submodule
+        if (!((CliGitAPIImpl)w.git).isAtLeastVersion(1,8,2,0)) {
+            System.err.println("git must be at least 1.8.2 to test 'submodule update --force'.");
+            return;
+        }
         w.cmd("git -C modules/ntp rm CHANGELOG");
 
         // perform a submodule update without force, and ensure that it does
